+++ app/policies/prep_kitchen/dashboard_policy.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module PrepKitchen
+  class DashboardPolicy < BasePolicy
+    def show?
+      prep_staff?
+    end
+  end
+end