+++ app/policies/prep_kitchen/recipe_policy.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module PrepKitchen
+  class RecipePolicy < BasePolicy
+    def index?
+      prep_manager?
+    end
+  end
+end