+++ app/policies/prep_kitchen/stop_list_policy.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module PrepKitchen
+  class StopListPolicy < BasePolicy
+    def index?
+      prep_staff?
+    end
+
+    def update?
+      prep_manager?
+    end
+  end
+end